// frontend/src/shared/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ConstructorStanding: {
            /** Position */
            position: string;
            /** Points */
            points: string;
            Constructor: components["schemas"]["ConstructorRecord"];
        } & {
            [key: string]: unknown;
        };
// ...
        DriverStanding: {
            /** Position */
            position: string;
            /** Points */
            points: string;
            Driver: components["schemas"]["DriverRecord"];
            /** Constructors */
            Constructors: components["schemas"]["ConstructorRecord"][];
        } & {
            [key: string]: unknown;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}